requirements.txt
django
djangorestframework
django-cors-headers
whitenoise
sendgrid
stripe
